+++ packages/input/src/MultiSelect.svelte
@@ -1,6 +1,6 @@
 <script>
 	import Select from './Select.svelte'
-	import { Pill } from '@svelte-spice/core'
+	import { Pill } from '@rokkit/core'
 
 	let className = ''
 	export { className as class }