+++ sites/stories/src/config/icons/index.js
@@ -1,4 +1,4 @@
-// import { iconStore } from '@svelte-spice/core'
+// import { iconStore } from '@rokkit/core'
 
 import NodeClosed from './NodeCollapsed.svelte'
 import NodeOpen from './NodeExpanded.svelte'