+++ packages/core/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @svelte-spice/core
+# @rokkit/core
+
+## 1.0.0-next.5
+
+### Patch Changes
+
+- Switched namespace for packages.
 
 ## 1.0.0-next.4